Move Swap Approve to new package
