+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swap/uniswap/UniswapFragment.kt
@@ -51,8 +51,6 @@ import io.horizontalsystems.bankwallet.modules.swap.SwapBaseFragment
 import io.horizontalsystems.bankwallet.modules.swap.SwapMainModule
 import io.horizontalsystems.bankwallet.modules.swap.SwapMainViewModel
 import io.horizontalsystems.bankwallet.modules.swap.allowance.SwapAllowanceViewModel
-import io.horizontalsystems.bankwallet.modules.swap.approve.SwapApproveModule
-import io.horizontalsystems.bankwallet.modules.swap.approve.confirmation.SwapApproveConfirmationModule
 import io.horizontalsystems.bankwallet.modules.swap.coincard.SwapCoinCardView
 import io.horizontalsystems.bankwallet.modules.swap.coincard.SwapCoinCardViewModel
 import io.horizontalsystems.bankwallet.modules.swap.confirmation.uniswap.UniswapConfirmationModule
@@ -64,6 +62,8 @@ import io.horizontalsystems.bankwallet.modules.swap.ui.SuggestionsBar
 import io.horizontalsystems.bankwallet.modules.swap.ui.SwapError
 import io.horizontalsystems.bankwallet.modules.swap.ui.SwitchCoinsSection
 import io.horizontalsystems.bankwallet.modules.swap.uniswap.UniswapTradeService.PriceImpactLevel
+import io.horizontalsystems.bankwallet.modules.swapx.approve.SwapApproveModule
+import io.horizontalsystems.bankwallet.modules.swapx.approve.confirmation.SwapApproveConfirmationModule
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 import io.horizontalsystems.bankwallet.ui.compose.Keyboard
 import io.horizontalsystems.bankwallet.ui.compose.components.TextImportantWarning
@@ -283,12 +283,12 @@ private fun UniswapScreen(
                             }
                         }
 
-                        viewModel.approveData?.let { data ->
-                            navController.slideFromBottom(
-                                R.id.swapApproveFragment,
-                                SwapApproveModule.prepareParams(data)
-                            )
-                        }
+//                        viewModel.approveData?.let { data ->
+//                            navController.slideFromBottom(
+//                                R.id.swapApproveFragment,
+//                                SwapApproveModule.prepareParams(data)
+//                            )
+//                        }
                     },
                     onTapProceed = {
                         viewModel.proceedParams?.let { sendEvmData ->
